requirements.txt
python-telegram-bot==20.5
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.32.0
apscheduler==3.10.4
